Use Worker for Cog HTTP API and respond like Replicate
This switches the old, synchronous predict API over to use the new
Worker code for running predictions in a subprocess so we're sharing a
common approach to running predictions between the redis worker code and
the Cog HTTP API.

At the same time, it defines prediction request and response types, and
uses these in the Cog HTTP API (annotating/extending them with the model
input and output types as appropriate). This is the first step in making
the Cog HTTP API more similar to Replicate's own API.

The change is intended to be backwards compatible.

Co-authored-by: Dominic Baggott <dominic.baggott@gmail.com>
Co-authored-by: Nick Stenning <nick@whiteink.com>
Signed-off-by: Dominic Baggott <dominic.baggott@gmail.com>
Signed-off-by: Nick Stenning <nick@whiteink.com>