WIP: Make (some) File inputs picklable
This commit makes it possible for the HTTP API to handle File input
types with the new Worker. Because Worker is on the other side of a
process boundary, whatever the inputs validate to must be picklable, and
file handles are not (nor will they ever be).

This demonstrates one way we might deal with this, by making the
"file-like" object we pass in a lazily-executing wrapper around the HTTP
response. This works, but does have some downsides:

- File and Path now behave differently: File inputs are downloaded in
  the predictor process, Path inputs are downloaded at validation-time.

- We now don't detect problems with the URL (e.g. it 404s) until we
  actually call predict().

Note that we'll need to implement a bunch more of the IOBase API on
URLFile if we want to take this path.

Signed-off-by: Nick Stenning <nick@whiteink.com>